backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1
pandas==2.1.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Azure Services
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0

# OpenAI (for future GPT integration)
openai==1.3.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For advanced features
# azure-cognitiveservices-speech==1.34.0  # For speech functionality
# azure-ai-textanalytics==5.3.0          # For text analytics
